organized code better
